[tutasdk] move id-related code to its own module
