+++ tuta-sdk/rust/sdk/src/blobs/blob_facade.rs
@@ -7,7 +7,6 @@ use crate::crypto::randomizer_facade::RandomizerFacade;
 use crate::entities::generated::storage::{BlobGetIn, BlobPostOut, BlobServerAccessInfo};
 use crate::entities::generated::sys::BlobReferenceTokenWrapper;
 use crate::entities::Entity;
-use crate::generated_id::GeneratedId;
 use crate::instance_mapper::InstanceMapper;
 use crate::json_element::RawEntity;
 use crate::json_serializer::JsonSerializer;
@@ -17,6 +16,7 @@ use crate::rest_client::{RestClientOptions, RestResponse};
 use crate::rest_error::HttpError;
 use crate::tutanota_constants::{ArchiveDataType, MAX_BLOB_SIZE_BYTES};
 use crate::type_model_provider::init_type_model_provider;
+use crate::GeneratedId;
 use crate::{crypto, ApiCallError, HeadersProvider};
 use base64::Engine;
 use crypto::sha256;
@@ -256,17 +256,17 @@ mod tests {
 	use crate::blobs::blob_access_token_facade::MockBlobAccessTokenFacade;
 	use crate::crypto::randomizer_facade::test_util::DeterministicRng;
 	use crate::crypto::randomizer_facade::RandomizerFacade;
-	use crate::custom_id::CustomId;
 	use crate::entities::generated::storage::BlobPostOut;
 	use crate::entities::generated::storage::{BlobServerAccessInfo, BlobServerUrl};
 	use crate::entities::generated::sys::BlobReferenceTokenWrapper;
-	use crate::generated_id::GeneratedId;
 	use crate::rest_client::MockRestClient;
 	use crate::rest_client::RestClientOptions;
 	use crate::rest_client::RestResponse;
 	use crate::tutanota_constants::ArchiveDataType;
 	use crate::type_model_provider::init_type_model_provider;
 	use crate::util::test_utils::create_test_entity;
+	use crate::CustomId;
+	use crate::GeneratedId;
 	use crate::GenericAesKey;
 	use crate::HeadersProvider;
 	use crate::InstanceMapper;
