+++ tuta-sdk/rust/sdk/src/crypto_entity_client.rs
@@ -5,10 +5,10 @@ use crate::entities::entity_facade::EntityFacade;
 use crate::entities::Entity;
 #[cfg_attr(test, mockall_double::double)]
 use crate::entity_client::EntityClient;
-use crate::entity_client::IdType;
-use crate::generated_id::GeneratedId;
+use crate::id::id_tuple::IdType;
 use crate::instance_mapper::InstanceMapper;
 use crate::metamodel::TypeModel;
+use crate::GeneratedId;
 use crate::{ApiCallError, ListLoadDirection};
 use serde::Deserialize;
 use std::sync::Arc;
