+++ tuta-sdk/rust/sdk/src/id/generated_id.rs
@@ -1,4 +1,4 @@
-use crate::entity_client::{BaseIdType, IdType};
+use crate::id::id_tuple::{BaseIdType, IdType};
 use serde::de::{Error, Visitor};
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::borrow::ToOwned;
