+++ tuta-sdk/rust/sdk/src/key_cache.rs
@@ -1,7 +1,7 @@
 use crate::crypto::Aes256Key;
 use crate::entities::generated::sys::User;
-use crate::generated_id::GeneratedId;
 use crate::key_loader_facade::VersionedAesKey;
+use crate::GeneratedId;
 use std::collections::HashMap;
 use std::sync::RwLock;
