+++ tuta-sdk/rust/sdk/src/key_loader_facade.rs
@@ -1,14 +1,15 @@
 use crate::crypto::key::{AsymmetricKeyPair, GenericAesKey, KeyLoadError};
 use crate::crypto::key_encryption::decrypt_key_pair;
-use crate::custom_id::CustomId;
 use crate::entities::generated::sys::{Group, GroupKey, KeyPair};
-use crate::generated_id::GeneratedId;
 #[cfg_attr(test, mockall_double::double)]
 use crate::typed_entity_client::TypedEntityClient;
 #[cfg_attr(test, mockall_double::double)]
 use crate::user_facade::UserFacade;
 use crate::util::Versioned;
-use crate::{IdTupleCustom, ListLoadDirection};
+use crate::CustomId;
+use crate::GeneratedId;
+use crate::IdTupleCustom;
+use crate::ListLoadDirection;
 use futures::future::BoxFuture;
 use std::cmp::Ordering;
 use std::sync::Arc;
@@ -27,7 +28,7 @@ impl KeyLoaderFacade {
 		}
 	}
 
-	/// Load the symmetric group key for the group_id with the provided requestedVersion.
+	/// Load the symmetric group key for the groupId with the provided requestedVersion.
 	/// `currentGroupKey` needs to be set if the user is not a member of the group (e.g. an admin)
 	pub async fn load_sym_group_key(
 		&self,
@@ -350,13 +351,13 @@ mod tests {
 	use crate::crypto::randomizer_facade::test_util::make_thread_rng_facade;
 	use crate::crypto::randomizer_facade::RandomizerFacade;
 	use crate::crypto::{aes::Iv, Aes256Key, PQKeyPairs};
-	use crate::custom_id::CustomId;
 	use crate::entities::generated::sys::{GroupKeysRef, GroupMembership, KeyPair};
 	use crate::key_cache::MockKeyCache;
 	use crate::typed_entity_client::MockTypedEntityClient;
 	use crate::user_facade::MockUserFacade;
 	use crate::util::get_vec_reversed;
 	use crate::util::test_utils::{generate_random_group, random_aes256_key};
+	use crate::CustomId;
 	use crate::{IdTupleCustom, IdTupleGenerated};
 	use mockall::predicate;
 	use std::array::from_fn;
